+++ docs/src/reference/sdemo/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## `v1.3.2`
 
+- ** fixed** the problem with hyper-parameters not being restored
+- ** fixed** the typing issue in Shapley values
+
+## `v1.3.2`
+
 - **fixed** a bug where the Shapley values calculation would not work when applied to a dataset not used for testing
 
 ## `v1.3.1`